started moving UI elements into own class called GhostAvatarControlsUI
